monononoke: bump treepack_part buffer size
Summary:
We've seen some requests takes many seconds to complete in non-master regions.
Turned out that happens when a request requires a lot of fetches from
manifold, but if all entries are in the cache then the latency from different
regions is roughly the same.

Increasing the buffer size helps because we do more manifold fetches in parallel.
10K was chosen quite randomly, I just checked that it works fine in different
regions.

There is a concern that increasing buffer size will overload manifold in
master region. However, I hope that manifold blobstore will help us with that.

Reviewed By: farnz

Differential Revision: D8610582

fbshipit-source-id: 900eac66141c3ab686024e34235a9ca7454a9033