fix a bug in the _checkforeignmodules() code
Summary:
Fix the `_collectimport()` code to more accurately figure out module paths.

When used from a statement like `import foo.bar` the `__import__()` function
returns the `foo` module.  However, if used like `from foo.bar import baz` it
returns the `foo.bar` module.  The `_collectimport()` code did not expect this
behavior and assumed that it always returned `foo`.  This resulted in it doing
the wrong thing when trying to resolve nested imports.

In most cases it would just fail with an attribute error and it would then
swallow this exceptions.  However in some cases it was able to find a
non-module object.  For instance, when processing the `mercurial.extensions`
module it would find the `extensions()` function inside this module.  Calling
`inspect.getabsfile()` on this function doesn't always produce the correct
result.  If loaded from a pre-compiled .pyc file the function may have a
relative path baked in as its function name.  The `inspect.getabsfile()` code
then simply prepends this with the current working directory, despite the fact
that this wasn't actually imported from the current directory.

I also changed the code to use inspect.getfile() rather than
inspect.getabsfile().  The getabsfile() function attempts to do other
normalization and string manipulation that we don't need.  (It tries to find
`*.py` names rather than `*.pyc` or `*.pyo` names, and it normalizes case.)

Reviewed By: ryanmce, quark-zju

Differential Revision: D8601652

fbshipit-source-id: db9be3c5cbbea83a880851036093aaa28b5b9b19