fix `metaedit` for non-continuous subsets of linear commits
Summary:
Assume the stack:
```
o - C
|
o - B
|
o - A
|
o - Root
```
then if we need to metaedit commits A and B the stack becomes:
```
o - C' (metaedited)
|  o - B'
|  |
|  o - A' (metaedited)
| /
o - Root
```
This bug affects also Jellyfish while submitting the stack of commits where some commit in the middle is already assotiated with Phab diff.

The bug is caused by the logic where we firstly metaedit and rebase the commits to metaedit in their topological order and only after all their descendants. So for now the given commits are processed with their descendants in topological order to track new parents for each commit.

Reviewed By: mitrandir77

Differential Revision: D8591897

fbshipit-source-id: 330a598e084af60cff42255510aca420bfe389de