treemanifest: prefetch public parents when receiving drafts
Summary:
When we apply draft commits to the repository, through
pull/unbundler/etc, we want to prefetch their public parent trees. If we don't,
when the draft commit is accessed it has the root node, but not some of the
children, and therefore attempts to download all the necessary trees ondemand
without any base trees. This is very expensive.

Let's prefetch the public parents, which should be relatively cheap since it can
delta against a nearby tree, and which is probably required anyway since
applying draft commits usually indicates they will be accessed soon.

Reviewed By: quark-zju

Differential Revision: D8654617

fbshipit-source-id: 63ee1e958fc0a0abdb8fa7bdb8c5d9e01181b6a7