sampling: avoid cyclic reference in atexit handler
Summary:
When debugging T30931730, I created a samll repo with almost all extensions
disabled. Then there is no leak report for the eden importer. Then I start
bisecting the extensions, and found the sampling extension is one of the
troublemakers. This is similar to a fix in hgsubverison [1].

This does not reproduce outside the eden importer because the atexit handlers
are consumed by `dispatch.py`. Regarding on whether eden importer should also
process atexit handlers or not remains a question. Anyway, avoiding cyclic
references shouldn't make things worse.

[1]: https://bitbucket.org/durin42/hgsubversion/commits/3c4fbf59e503d3812eb923e2b85a0121cc71657e

Reviewed By: chadaustin

Differential Revision: D8675401

fbshipit-source-id: cf87869d2006d5166c697c2a86b65ece6cbe8c06