record: remove name encoded backups
Summary:
Previously we were encoding the full filepath into a single filename to
use as a backup file. If the full path was over 255 characters, then we couldn't
create a file with the name.

Since the user never sees this name unless they run --debug, let's just get rid
of the encoded name and use a purely temporary name.

Reviewed By: farnz

Differential Revision: D8695749

fbshipit-source-id: 18b6d00e5f4c856b7dd6073e9d8c9677c4baea69