clone: close repo during clone
Summary:
A future diff will make us depend on repo.close() being called to
finalize the writing of some tree data. It turns out we didn't call close on the
destination repo during a clone, which caused some tests to change. Let's call
close on that repo to prevent a test change in the next diff.

Reviewed By: phillco

Differential Revision: D8670680

fbshipit-source-id: ad63ccfc51ec4c3de2dad494fdb1ae8f6ba8e40a