mononoke: fix test-gettreepack.t
Summary:
New parameter depth was added. Mononoke just ignores, which if correct, but can
be less efficient. However test-gettreepack.t was failing because it didn't
supply the parameter. This diff fixes it by sending a big depth to the server

P. S.
Attached task tracks the support of depth parameter on Mononoke server-side

Reviewed By: lukaspiatkowski

Differential Revision: D8712906

fbshipit-source-id: 8246344b3fa39f00eb39f2262b4aa5aa74834be2