remotefilelog, treemanifest: add test demonstrating excessive gettreepacks on amend
Summary:
When amending a commit containing a file rename for a file with a long history
where the original content was usually on a separate branch, we end up in
`remotefilelogctx._adjustlinknode` for every commit the file was modified in.
Each of these looks up the manifest for the file, which causes a tree fetch.

Reviewed By: quark-zju

Differential Revision: D8691179

fbshipit-source-id: 088a6b402a79d37790ce7ac4481a39c657f690ee