templater: allow non-ASCII fillchar in pad()
Summary:
Hg's `pad()` template function can only fill with a single character. To assert this, Hg checks the length of the fill string. Because the fill string is a Python 2 `str`, Hg rejects UTF-8 `'☃'` (`'\xe2\x98\x83'`) because its length is 3.

Instead of checking the length of the fill string, check the width of the fill string while taking encoding into account. This lets us use ☃ as a pad fill character.

For ASCII fill strings in `pad()`, this diff should not change behaviour.

Reviewed By: phillco

Differential Revision: D8730727

fbshipit-source-id: ad015aa4cae274160bc2726e0cb5e570e6edcc1e