revisionstore: change delta base to be an Option<>
Summary:
Deltas may not have bases if they are a full text. Let's represent
that as an Option instead of as a magical null id value. This has the nice
effect of moving the decision to serialize a missing delta base down into the
serializer instead of up at the delta chain construction level.

Reviewed By: quark-zju

Differential Revision: D8739231

fbshipit-source-id: b58bd40dae45cb85890812db21e7eeff46aa6b4e