sampling: fix sampling for keys starting with e, k, y
Summary:
`lstrip` removes *all* of the characters in the given set from the beginning of the string. It doesn't just remove the given prefix, which I think was the intention.

So if you specified `key.env_vars=...`, `key.e` would be stripped, not just `key.`, and `nv_vars` would be set instead of `env_vars`

Reviewed By: quark-zju

Differential Revision: D8810694

fbshipit-source-id: 5f93bd9bdb660430c59ed7ae7e6ff7a7edcf4764