fastmanifest: switch to using manifestlog.ui where possible
Summary:
Previously, the fastmanifest factory class kept a copy of the initial ui object
from uisetup(). This meant it executed all things using the configuration of the
first repository. In a command that opens many repositories, like hg pull or hg
gc, this caused it to use the wrong configuration.

To fix this, let's switch as much of the fastmanifest factory to use
manifestlog.ui (which is the repo specific ui) instead of self.ui. We can't get
rid of self.ui just yet, since it's still required for the revlog.add()
override, but this fixes the hg gc breakage. We should fix the revlog.add case
later too.

Reviewed By: quark-zju

Differential Revision: D8840003

fbshipit-source-id: 037dcf98422826a0223a9941c2a5931b7408b2a5