Add second-order diff display for changes since last Phabricator upload
Summary:
The `arcdiff` hg extension is useful in concept - showing the changes since last Phabricator upload, but in practice there is usually too much noise in the output from rebasing. For example, consider the case where I have a stack and haven't re-uploaded the current commit. If one of the ancestors in the stack has changed and I use `--changes-since-last-arc-diff` I end up seeing those changes mixed plus the changes to the the ancestor.

Here we introduce a feature to show only a diff of the current commit vs. Phabriactor's version of the commit i.e. a diff of a diff, or "second-order" diff. The goal is not to perfectly reconstruct the changes, just to highlight what changed for human consumption. There are no line-ranges or context in the diff output, but it still seems pretty usable.

This new feature is invoked via the option `--changes-since-last-arc-diff-2o` to `hg diff`.

Reviewed By: quark-zju

Differential Revision: D8787504

fbshipit-source-id: 2885a118a712a3d9187e6fce67a465e7b664c3ca