commit cloud: replace default path when pulling from Mononoke
Summary:
To make Mononoke dogfoodable, we need to add this change.
In D8734958 I've added force pulling from `infinitepush` path in some cases
(i.e. when user does `hg pull -r HASH`)

However that's not enough. Changing the path to `infinitepush` results in
remotenames adding lots of new bookmarks to `.hg/remotenames` file like
`infinitepush/master`, `infinitepush/stable` and so on. First of all, that means that
they will show up in smartlog and make it more cluttered. Secondly, and more
importantly, it will result in adding lots of bookmarks to the repo.journal -
https://fburl.com/80hy1rv5 . Turns out, that can be very slow.

Reviewed By: farnz

Differential Revision: D8859001

fbshipit-source-id: 973c0ab13563a7010077fdb016f43106b1fba744