test-fb-hgext-p4fastimport-seqimport-reorg: avoid chaining p4 commands using pipes
Summary:
Recently there is a high (30%+) chance that the linux build worker times out.
I tracked it down to this test. It hangs forever on a couple of workers.
The process tree looks like:

  python,1904510 run-tests.py ...
    |-bash,2067768 .../test-fb-hgext-p4fastimport-seqimport-reorg.t.sh
    |   |-p4,2073159 change -o 3
    |   |-p4,2073161 change -fi
    |   |-p4d,2067785 -f -J off
    |   `-sed,2073160 s: foo: foo IMPORTER_IGNORE_REORG@:
    |-{python},2067733
    `-{python},2067780

It seems like a deadlock or race condition somewhere. So let's break the long
command to make sure `change -fi` runs after `change -o 3` completes.

Reviewed By: DurhamG

Differential Revision: D8895528

fbshipit-source-id: 504a94943bc31c8c137df65b34b52207de6766f0