rebase: fix fallback if merge.update leaves files unresolved, but doesn't raise InMemoryMergeConflictsError
Summary:
As written, merge tools and drivers are responsible for raising `InMemoryMergeConflictsError` themselves. We shouldn't finish merge.update with unresolved files under IMM.

However, we had this fallback in case we missed a spot, and it turns out there was such a spot: D8886237.

D8699551's addition of required parameters to the constructor broke this fallback. So, this adds the parameters back in.

Reviewed By: quark-zju

Differential Revision: D8886238

fbshipit-source-id: 111fcce6a5a38d126c0118208e508c04b7566971