merge: properly raise a InMemoryMergeConflictsError if driver raises an exception
Summary:
Merge drivers can raise exceptions. If they do, we need to restart the rebase without IMM in case they might work without IMM.

Before, we just exited out of merge.update with a positive "unresolved" count. This change raises a proper exception so we don't hit the fallback in D8886238, and have access to the failed files. It also rewrites some logic that was murky.

I think a better approach overall might be to not have driver raise exceptions, and cancel the rebase in that case or otherwise treat that case more harshly. We'd need to change existing drivers in that case. Or we can just handle this case more explicitly by adding a new `type` to `InMemoryMergeConflictsError`.

 ---

Note that because of the nice fallback in D8886238, all this does is change the line:

```
hit merge conflicts (in 1 file); switching to on-disk merge
```

to

```
hit merge conflicts (in FILE); switching to on-disk merge
```

but it's important to have explicit flow control so we can make future changes with confidence.

 ---

Reviewed By: quark-zju

Differential Revision: D8886237

fbshipit-source-id: 4e10e960f7a8d8f37de177d0596d2f3da16be4d9