Fix incorrect base-rev and file lists in hg diff --since-last-submit-2o
Summary:
When creating the two first-order diffs we are currently diffing the two revs to the parent of rev1. This was something I was playing around with before but actually doesn't make sense. This returns things to diffing rev1 to rev1's parent and rev2 to rev2's parent.

I think the problem I was trying to work around was seeing files appearing to be changed not to do with either rev1 or rev2. The correct fix for this is to only consider files which are in rev1 or rev2.

Reviewed By: quark-zju

Differential Revision: D8904312

fbshipit-source-id: f48bd498cc27ea650a859d7b318f32520f946150