admin: add config fbpkg subcommand
Summary:
This command will create a PROD bookmark on the currently checked out commit of the choosen mononoke-config repo, then import it into a temp directory and finally invoke fbpkg build to create a package with that imported content.
It also calls fbpkg build inside the mononoke-config repo so that the VCS info from mercurial is persisted in fbpkg metadata.

Reviewed By: farnz

Differential Revision: D8907916

fbshipit-source-id: ee1843bf23c9bffe524d2b20c7322f84825e2663