Initial structure for a SkiplistIndex that can be put into the API server.
Summary:
The initial structure for storing skip list edges for the commit graph, to support faster reachability queries.

The plan for this index is to have two main methods:
- lazy_index, which takes a starting node, and a max_depth value, and indexes all nodes within distance max_depth of the start, or until it reaches a node which has already been indexed.
- query_with_indexing, which attempts to answer a reachability query, while indexing nodes it comes across with lazy_index if necessary.

Initialization of the index will start with a warm up by calling lazy_index on master, with a choice of how deep to warm up. 240K seems to be a good value, because the average query we've seen has distance <= 80K, and it is still a reasonable warm up time (4 minutes to obtain all database information if you assume 1ms per get_parents call, but even less if this information has already been cached by another warm up).

Then, when a query comes in, it would perform as follows:
- Pursue as many skip list edges as possible, assuming we are starting at an indexed node, until we reach a node which we can't pursue skip edges from, either because it is unindexed, or because it is a merge node.
- Once at a node like this:
    - If it is unindexed, perform a lazy_index call, where the depth parameter could either be a magic number, or it could be related to the query. For example, the difference between the unindexed node's generation and the destination generation. Then restart the query from this node.
    - If it is indexed but doesnt have skip edges, then a future must be spawned for each parent, as another query_with_indexing call. Then the query waits for all futures to complete, and returns the or of their results. This gives us the benefit of doing the indexing work.

The most noteworthy part of this initial structure is the choice of chashmap. This is a good choice for this structure because:
- After a node's skip edges are inserted, they are immutable. Thus we have low contention for writes on the same key.
- It separates locks on key buckets. So one thread can perform a reachability query while other threads are updating the index for unrelated commits in parallel.
- Even if two threads try to update the same key, for example if two different queries are simultaneously filling out the index, it doesn't
  matter which write wins out. In our case, the skip edges are deterministic, so the writes will be identical. Even if they were randomized,
  either write choice will maintain consistency of the index.

Reviewed By: StanislavGlebik

Differential Revision: D8919861

fbshipit-source-id: b6d0d61fb5484d406fec269d15b3e6c4eb9dac4a