rebase: improve the IMM working copy check
Summary:
I noticed when `amend --rebase`ing that rebase was refusing to run because my working copy was dirty. That shouldn't happen with `amend --rebase` because the working copy won't move.

I realized you can pass revisions in the rebaseset that won't be rebased, which is what was happening here. So, let's move this check to look at the generated `destmap`, which is the definitive source of truth for what will move.

Reviewed By: quark-zju

Differential Revision: D8921695

fbshipit-source-id: 292e8d14342adf70a4f693d91a9a3f2aad4f2ff9